--- {{cookiecutter.project_dashed}}-src/{{cookiecutter.project_namespace}}/forms.py
@@ -1,5 +1,0 @@
-from keg_auth.forms import user_form as user_form_base
-
-
-def user_form(config, allow_superuser, endpoint, fields=['name', 'is_enabled']):
-    return user_form_base(config, allow_superuser, endpoint, fields=fields)
