--- {{cookiecutter.project_dashed}}-src/{{cookiecutter.project_namespace}}/views/__init__.py
@@ -1,5 +1,0 @@
-from .auth import auth_bp
-from .private import private_bp
-from .public import public_bp
-
-blueprints = (auth_bp, public_bp, private_bp,)
