--- {{cookiecutter.project_dashed}}-src/{{cookiecutter.project_namespace}}/views/auth.py
@@ -1,18 +1,0 @@
-import logging
-
-from keg_auth import make_blueprint
-from keg_auth.views import User as UserBase
-
-from {{cookiecutter.project_namespace}} import forms
-from {{cookiecutter.project_namespace}}.extensions import auth_manager
-
-log = logging.getLogger(__name__)
-
-
-class User(UserBase):
-    # need to make this a static method so it isn't bound on the view instance
-    form_cls = staticmethod(forms.user_form)
-
-
-# This blueprint is for keg-auth's views (Login, user management, etc.)
-auth_bp = make_blueprint(__name__, auth_manager, user_crud_cls=User)
